pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ievv-auth"
dynamic = ["version"]
description = "Authentication modules for the Django framework."
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    { name = "Appresso developers", email = "post@appresso.no" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Intended Audience :: Developers",
    "License :: OSI Approved",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
]

dependencies = [
    "cryptography==2.8",
    "Django>=4.0.0,<5.0.0",
    "django-ipware==2.1.0",
    "djangorestframework",
    "ievv_opensource>=9.2.2,<10.0.0",
    "psycopg2",
    "PyJWT==2.8.0",
    "python-dateutil",
]

[project.optional-dependencies]
dev = []

test = [
    'model-bakery',
    "pytest",
    "pytest-django",
    "pytest-env",
]

[project.urls]
Homepage = "https://github.com/appressoas/ievv_auth"

[tool.hatch.version]
path = "ievv_auth/__init__.py"

[tool.hatch.build]
packages = [
    "/ievv_auth",
]
exclude = []

[tool.commitizen]
name = "cz_conventional_commits"
version = "3.1.2"
version_files = [
    "ievv_auth/__init__.py:__version__"
]
tag_format = "$version"
update_changelog_on_bump = true
